single water

+++ pycommon/dir_job.py
@@ -3,6 +3,7 @@
 import argparse
 import os
 import re
+import sys
 from common import *
 
 def run_commands(coms):
@@ -10,7 +11,7 @@ def run_commands(coms):
         os.system(com)
     return
 
-def reset(package_name, dsave, save_files, Lrun):
+def reset(package_name,command, dsave, save_files, Lrun):
 
     commands=[]
     if package_name == 'amp':
@@ -25,16 +26,18 @@ def reset(package_name, dsave, save_files, Lrun):
             print("first save files to %s manually" % dsave)
             return
         '''
-        com = 'rm * '
-        print(com)
-        commands.append(com)
-        com = 'rm -r *ampdb'
-        print(com)
-        commands.append(com)
-        com = 'cp %s/* .' % dsave
-        print(com)
-        commands.append(com)
-
+        if command == 'reset':
+            com = 'rm * '
+            print(com)
+            commands.append(com)
+            com = 'rm -r *ampdb'
+            print(com)
+            commands.append(com)
+            com = 'cp %s/* .' % dsave
+            print(com)
+            commands.append(com)
+        elif command == 'mkdir':
+            
         if Lrun or yes_or_no("Did you save and will you run?"):
             run_commands(commands)
             
@@ -43,13 +46,19 @@ def reset(package_name, dsave, save_files, Lrun):
 
 def main():
     parser = argparse.ArgumentParser(description='reset directory with saving mode')
-    parser.add_argument('package', choices=['amp'], help='choose packagename')
-    parser.add_argument('-s', '--saving_dir', default='ini', help='save files to ./ini')
+    parser.add_argument('-j', '--job', choices=['amp'], help='choose packagename')
+    parser.add_argument('-c', '--command', choices=['reset','mkdir'], help='command=="mkdir": make directory and copy basic files\
+    \n\t\t=="reset": save files and rerun at the same directory')
+    parser.add_argument('-s', '--saving_dir', help='save files to ./dir')
     parser.add_argument('-f', '--saving_files', nargs='*', help='select files to be saved')
     parser.add_argument('-r', '--run', action='store_true', help='run command if not just show')
     args = parser.parse_args()
 
-    reset(args.package, args.saving_dir, args.saving_files, args.run)
+    if not args.job:
+        print("input -j job such as 'amp'")
+        sys.exit(0)
+
+    reset(args.job, args.command, args.saving_dir, args.saving_files, args.run)
     return 0
 
 if __name__ == '__main__':
